Restructure examples directory to support multiple examples
