--- example/style.css
@@ -1,59 +1,0 @@
-:root {
-  color-scheme: light dark;
-}
-
-body {
-  display: flex;
-  margin: 0;
-  padding: 2em;
-  box-sizing: border-box;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
-  font-family: sans-serif;
-}
-
-form {
-  min-width: 300px;
-}
-
-form > label {
-  display: flex;
-  gap: .5em;
-  margin-block: 1em .2em;
-}
-
-input:not([type=checkbox]):not([type=radio]), textarea, select {
-  font: inherit;
-  padding: .4em .5em;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-button {
-  margin-block: 2em;
-  width: 100%;
-  font: inherit;
-  padding: .5em;
-}
-
-.error {
-  font-size: .8em;
-  color: darkred;
-}
-
-@media (prefers-color-scheme: dark) {
-  .error {
-    color: lightcoral;
-  }
-}
-
-output {
-  font-family: monospace;
-  white-space: pre;
-}
-
-output div {
-  margin-block: .5em;
-}
