+++ index.html
@@ -3,11 +3,11 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="./example/style.css">
-    <title>React Zoom Form Example</title>
+    <link rel="stylesheet" href="./examples/style.css">
+    <title>React Zoom Form Examples</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./example/index.tsx"></script>
+    <script type="module" src="./examples/index.tsx"></script>
   </body>
 </html>
