+++ package.json
@@ -55,6 +55,7 @@
     "vite": "^4.4.2",
     "vite-plugin-dts": "^3.2.0",
     "vitest": "^0.33.0",
+    "wouter": "^2.11.0",
     "zod": "^3.21.4"
   }
 }
